pyproject.toml
[project]
name = "a2a-server-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "a2a-sdk>=0.3.10",
    "fastapi>=0.121.0",
    "fastmcp>=2.13.0.2",
    "google-adk>=1.18.0",
    "nest-asyncio>=1.6.0",
    "openai>=2.7.1",
    "strands-agents>=1.15.0",
    "strands-agents-tools[a2a-client]>=0.2.14",
    "uvicorn>=0.38.0",
]
